+++ chain/src/interfaces/ICapTable.sol
@@ -1,7 +1,9 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-import { AccessControlDefaultAdminRules } from "openzeppelin/contracts/access/AccessControlDefaultAdminRules.sol";
+import { IAccessControlDefaultAdminRules } from "openzeppelin-contracts/contracts/access/IAccessControlDefaultAdminRules.sol";
+
+import { AccessControlDefaultAdminRulesUpgradeable } from "openzeppelin-contracts-upgradeable/contracts/access/AccessControlDefaultAdminRulesUpgradeable.sol";
 import { Issuer, Stakeholder, StockClass, ActivePositions, SecIdsStockClass, StockLegendTemplate, InitialShares, ShareNumbersIssued, StockParams, StockParamsQuantity, StockIssuanceParams } from "../lib/Structs.sol";
 
 interface ICapTable {
@@ -75,11 +77,11 @@ interface ICapTable {
     function getTotalActiveSecuritiesCount() external view returns (uint256);
 
     // Function to get the timestamp of an active position
-    function getActivePosition(bytes16 stakeholderId, bytes16 securityId) external view returns (bytes16, uint, uint, uint40);
+    function getActivePosition(bytes16 stakeholderId, bytes16 securityId) external view returns (bytes16, uint256, uint256, uint40);
 
     /// @notice Get the avg active position for the stakeholder by dividing the first return value (quantityPrice) by the second (quantity)
     ///  the timestamp is the time of the latest position
-    function getAveragePosition(bytes16 stakeholderId, bytes16 stockClassId) external view returns (uint, uint, uint40);
+    function getAveragePosition(bytes16 stakeholderId, bytes16 stockClassId) external view returns (uint256, uint256, uint40);
 
     function issueStock(StockIssuanceParams calldata params) external;
